auth user endpoint
Got a lot of stuff left beforewe can use this endpoint. TODO: 1. Make sure that jwt works properly and ensures that only logged in user can interact with the platform. 2. I am reading the fastapi docs for Oauth2 related to this, the docs use a timer which essentially means that a user will be logged out after a certain amount of times, perhaps we could use cookies instead so that user will remain logged in until they wipe their browser history. 3. also adding some features in the MYSQLHandler for confirming if a user is verified or not (if said user have registered using an email address)